Add a get_from_entities()
@synack points out correctly that it is actually useful to join pools
and, say, datacenters. Or pools and servers that contain more servers
(theoretically a hypervisor). So I am renaming get_from_pools() to
get_from_entities(), which is capable of lax joining entities, and I am
creating a simple get_from_pools() that is strict about the type being
passed to get_from_entities()